pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.1",
    "attrs==25.3.0",
    "django==5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter==25.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.28.0",
    "inflection==0.5.1",
    "ipython>=9.4.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "markdown==3.8.2",
    "pillow>=11.3.0",
    "pyjwt==2.10.1",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "rpds-py==0.27.0",
    "sqlparse==0.5.3",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
]
